+++ site/src/components/Icons/CloseIcon.tsx
